Refactor: Create pydoctor.linker module to hold the docstring linker code, as well as the taglink() function.
Since the taglink() function is used basically everywhere, I've created an alias to keep the diff small.

Add the option to pass a class name as string in System.objectsOfType().

Minor modification to pydoctor.utils.findClassFromDottedName() to raise ValueError() instead of exiting (because it's now used in System.objectsOfType()).
